+++ SpaceCab/SpaceCab_Intro.ino
@@ -20,7 +20,7 @@ void titleScreen()
   int16_t logoYVal = 6 + level.yOffset.getInteger();  
   
   launchCustomer();
-  Rect playerRect = {player.getXDisplay(), player.getYDisplay(), PLAYER_WIDTH, PLAYER_HEIGHT};
+  Rect playerRect = {static_cast<int16_t>(player.getXDisplay()), static_cast<int16_t>(player.getYDisplay()), PLAYER_WIDTH, PLAYER_HEIGHT};
   Rect customerRect = {customerXVal, customerYVal, CUSTOMER_WIDTH, CUSTOMER_HEIGHT};
   handleInput();
   drawLevel();
